make private classes private, dont pass params

+++ manifests/cli.pp
@@ -4,18 +4,21 @@
 #
 class jenkins::cli {
 
+  if $caller_module_name != $module_name {
+    fail("Use of private class ${name} by ${caller_module_name}")
+  }
+
   $jar = '/usr/lib/jenkins/jenkins-cli.jar'
   $extract_jar = 'unzip /usr/lib/jenkins/jenkins.war WEB-INF/jenkins-cli.jar'
   $move_jar = "mv WEB-INF/jenkins-cli.jar ${jar}"
   $remove_dir = 'rm -rf WEB-INF'
 
-  exec {
-    'jenkins-cli' :
-      command => "${extract_jar} && ${move_jar} && ${remove_dir}",
-      path    => ['/bin', '/usr/bin'],
-      cwd     => '/tmp',
-      creates => $jar,
-      require => Package['jenkins'];
+  exec { 'jenkins-cli' :
+    command => "${extract_jar} && ${move_jar} && ${remove_dir}",
+    path    => ['/bin', '/usr/bin'],
+    cwd     => '/tmp',
+    creates => $jar,
+    require => Package['jenkins'],
   }
 
 }
